requirements.txt
# Sentiment Analysis Chatbot - Requirements
#
# This project uses only Python standard library
# No external dependencies required!
#
# Python Version: 3.8+
#
# For development/testing (optional):
# pytest>=7.0.0
# black>=22.0.0
# flake8>=4.0.0
#
# Note: The core application runs without any pip installs
# API Requirements (Optional - for REST API feature)
flask==3.0.0
flask-cors==4.0.0
# AI Enhancement (Optional - for conversational AI responses)
google-generativeai>=0.3.0
python-dotenv>=1.0.0
